chore(e2e): externalizes/moves client.waitForVisit
